requirements.txt
fastapi==0.109.0
uvicorn[standard]==0.27.0
pydantic
pydantic-settings
python-dotenv==1.0.0
python-multipart==0.0.6
pytest==7.4.3
httpx==0.25.2
pytest-asyncio==0.21.1
uvicorn[standard]==0.27.0